Update countdown.js
the old method of constantly incrementing seconds and dividing that number by some randomly generated interval fails to change the second value on the counter if the seconds counter gets too big.

attempting to fix that by checking if the seconds counter gets to 61, and if it does reset it to a value between 0 and 59.  else add one second to the seconds counter.